+++ packages/extras/.storybook/Container.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { ThemeProvider } from '@bubbles-ui/components';
+
+export const Container = ({ story, id, context }) => {
+  return (
+    <React.StrictMode>
+      <ThemeProvider>
+        <div className={id.toLowerCase()} data-floating-menu-container role="main">
+          {story()}
+        </div>
+      </ThemeProvider>
+    </React.StrictMode>
+  );
+};