+++ packages/components/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @bubbles-ui/components
 
+## 1.2.46
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/icons@1.2.46
+
+## 1.2.45
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/icons@1.2.45
+
+## 1.2.44
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/icons@1.2.44
+
+## 1.2.43
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/icons@1.2.43
+
+## 1.2.42
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/icons@1.2.42
+
+## 1.2.41
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/icons@1.2.41
+
 ## 1.2.40
 
 ### Patch Changes