+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @bubbles-ui/icons
 
+## 1.2.46
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.45
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.44
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.43
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.42
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.41
+
+### Patch Changes
+
+- Versions bump
+
 ## 1.2.40
 
 ### Patch Changes