+++ packages/notifications/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @bubbles-ui/notifications
 
+## 1.2.46
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.46
+  - @bubbles-ui/icons@1.2.46
+
+## 1.2.45
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.45
+  - @bubbles-ui/icons@1.2.45
+
+## 1.2.44
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.44
+  - @bubbles-ui/icons@1.2.44
+
+## 1.2.43
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.43
+  - @bubbles-ui/icons@1.2.43
+
+## 1.2.42
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.42
+  - @bubbles-ui/icons@1.2.42
+
+## 1.2.41
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.41
+  - @bubbles-ui/icons@1.2.41
+
 ## 1.2.40
 
 ### Patch Changes